фикс учёта заявок

+++ core/App/HelpDesk/Ticket/DeleteCommand.php
@@ -2,8 +2,10 @@
 
 namespace Core\App\HelpDesk\Ticket;
 
+use Core\Domains\HelpDesk\Enums\TicketStatusEnum;
 use Core\Domains\HelpDesk\Models\TicketEntity;
 use Core\Domains\HelpDesk\Services\TicketService;
+use RuntimeException;
 
 readonly class DeleteCommand
 {
@@ -15,6 +17,10 @@ readonly class DeleteCommand
 
     public function execute(TicketEntity $ticket): void
     {
+        if ($ticket->getStatus() === TicketStatusEnum::CLOSED || $ticket->getStatus() === TicketStatusEnum::REJECTED) {
+            throw new RuntimeException('Нельзя удалить закрытую или отклонённую заявку');
+        }
+
         $this->ticketService->deleteById($ticket->getId());
     }
 }
